attempt to use entirely custom heat network templates.
these are derived from those of infrared, those that are on an undercloud,
some docs, and some gerrits.    we will try to have two nics, one acts
like the single-nic-vlans nic, the other nic is just extra and is there
to satisfy infrared's need to have an "external" nic and also makes it
easier to ssh to the overcloud nodes.